refactor package name import
refactor import org.guanzon.cas.model.inventory to org.guanzon.cas.inventory.models for all affected classes
