+++ src/main/java/org/guanzon/cas/inventory/base/Inventory.java
@@ -17,8 +17,8 @@ import org.guanzon.appdriver.constant.EditMode;
 import org.guanzon.appdriver.constant.TransactionStatus;
 import org.guanzon.appdriver.constant.UserRight;
 import org.guanzon.appdriver.iface.GRecord;
-import org.guanzon.cas.model.inventory.Model_Inventory;
-import org.guanzon.cas.model.inventory.Model_Inventory_Sub_Unit;
+import org.guanzon.cas.inventory.models.Model_Inventory;
+import org.guanzon.cas.inventory.models.Model_Inventory_Sub_Unit;
 import org.guanzon.cas.parameters.Brand;
 import org.guanzon.cas.parameters.Category;
 import org.guanzon.cas.parameters.Category_Level2;
